Create AdvancedPersistentRetroactiveQueue.py
When the modifications (mods) array of a node becomes full we create a new node to store the latest data. The back pointers of the new node should point to the previous (full) node. 

Ps: I am not sure this implementation is perfect but it was challenging because I had an improper pointer traversal and the mod logic made things complex.